requirements.txt
faiss_cpu==1.8.0
langchain==0.2.3
PyPDF2==3.0.1
sentence_transformers==3.0.0
streamlit==1.35.0
torch==2.3.0
transformers==4.41.0
sentencepiece==0.2.0
accelerate==0.31.0
bitsandbytes==0.43.1
